requirements.txt
pandas==1.0.1
google-cloud-bigquery==1.24.0
pandas-gbq==0.13.2
fastavro==0.23.6
pyarrow==0.17.1
tqdm==4.42.1
google-cloud-bigquery-storage==1.0.0